HTML5 Boilerplate
In the past for a long time, we are accustomed to replace the image with a negative text-indent method. In the latest HTML5 Boilerplate, I saw a new image replacement techniques. It uses CSS pseudo-elements, replacement techniques relative to the previous picture, it is a good solution to some mobile browser bugs, and no property covered, and does not require additional labeling.